optimize Versioned.get with a cache
From the original paper:
To further speed up the get operation, we maintain a cache that contains the version and corresponding index of the last read or write to this object. It is implemented as a 32 bit word that contains a version number in the lower 16 bits, and an index in the upper 16 bits. By keeping it the size of a word, we can atomically read and write this cache without using locks.